requirements.txt
requests
python-dotenv
matplotlib
Pillow